Moved and renamed header files of rotations
