+++ CMakeLists.txt
@@ -54,13 +54,13 @@ add_library(kindr_impl
 if(CATKIN_ENABLE_TESTING)
   catkin_add_gtest(test_${PROJECT_NAME}
     test/test_main.cpp
-	test/rotations/eigen/RotationMatrixTest.cpp
-	test/rotations/eigen/RotationQuaternionTest.cpp
-	test/rotations/eigen/RotationVectorTest.cpp
-	test/rotations/eigen/AngleAxisTest.cpp
-	test/rotations/eigen/EulerAnglesZyxTest.cpp
-	test/rotations/eigen/EulerAnglesXyzTest.cpp
-	test/rotations/RotationEigenTest.cpp
+	test/rotations/RotationMatrixTest.cpp
+	test/rotations/RotationQuaternionTest.cpp
+	test/rotations/RotationVectorTest.cpp
+	test/rotations/AngleAxisTest.cpp
+	test/rotations/EulerAnglesZyxTest.cpp
+	test/rotations/EulerAnglesXyzTest.cpp
+	test/rotations/RotationTest.cpp
   )
   if(TARGET test_${PROJECT_NAME})
     target_link_libraries(test_${PROJECT_NAME}
