+++ include/kindr/rotations/EulerAnglesZyxDiff.hpp
@@ -34,7 +34,7 @@
 #include "kindr/common/assert_macros.hpp"
 #include "kindr/common/assert_macros_eigen.hpp"
 #include "kindr/rotations/RotationDiffBase.hpp"
-#include "kindr/rotations/RotationEigen.hpp"
+#include "kindr/rotations/Rotation.hpp"
 #include "kindr/linear_algebra/LinearAlgebra.hpp"
 
 namespace kindr {
