+++ include/kindr/rotations/RotationMatrix.hpp
