+++ include/kindr/rotations/RotationQuaternion.hpp
