+++ include/kindr/rotations/RotationVector.hpp
