+++ src/kindr_impl.cpp
@@ -1,0 +1,21 @@
+/*
+ * kindr_impl.cpp
+ *
+ *  Created on: Jun 13, 2016
+ *      Author: gech
+ */
+
+
+
+
+#include "kindr/rotations/Rotation.hpp"
+
+
+int main(int argc, char **argv) {
+  kindr::RotationMatrixD rotaitonMatrix;
+  kindr::RotationVectorD rotationVector;
+  kindr::RotationQuaternionD rotationQuaternion;
+  kindr::AngleAxisD angleAxis;
+  kindr::EulerAnglesZyxD eulerAnglesZyx;
+  kindr::EulerAnglesXyzD eulerAnglesXyz;
+}
