+++ test/rotations/LocalAngularVelocityTest.cpp
