+++ test/rotations/RotationQuaternionDiffTest.cpp
