+++ test/rotations/RotationTest.cpp
@@ -28,8 +28,9 @@
 
 #include "kindr/common/gtest_eigen.hpp"
 #include "kindr/quaternions/QuaternionEigen.hpp"
-#include "kindr/rotations/RotationEigen.hpp"
 #include <string>
+
+#include "../../include/kindr/rotations/Rotation.hpp"
 #include "kindr/common/assert_macros_eigen.hpp"
